+++ demisto_sdk/commands/openapi_codegen/tests/openapi_codegen_test.py
@@ -32,8 +32,7 @@ class TestOpenAPICodeGen:
     test_files_path = os.path.join(git_path(), 'demisto_sdk', 'tests', 'test_files')
     swagger_path = os.path.join(test_files_path, 'swagger_pets.json')
 
-    def init_integration(self):
-        base_name = 'TestSwagger'
+    def init_integration(self, base_name: str = 'TestSwagger'):
         integration = OpenAPIIntegration(self.swagger_path, base_name,
                                          '-'.join(base_name.split(' ')).lower(),
                                          base_name.replace(' ', ''),
@@ -199,3 +198,59 @@ class TestOpenAPICodeGen:
         assert [c for c in integration.configuration['commands'] if c['name'] == 'post-pet-upload-image'][0]
         assert [c for c in integration.configuration['commands'] if c['name'] ==
                 'post-pet-upload-image-by-uploadimage'][0]
+
+    def test_file_not_overwritten(self):
+        """
+        Given:
+        - Configurations
+
+        When:
+        - Saving configuration file
+
+        Then:
+        - Ensure file does not overwrite given JSON file for open API code gen command.
+        """
+        integration = self.init_integration(base_name='swagger_pets')
+        with open(self.swagger_path, 'r') as f:
+            file_data_before_config_save = json.loads(f.read())
+        integration.save_config(integration.configuration, self.test_files_path)
+        with open(self.swagger_path, 'r') as f:
+            file_data_after_config_save = json.loads(f.read())
+        assert file_data_after_config_save == file_data_before_config_save
+        os.remove(os.path.join(self.test_files_path, f'{integration.base_name}_config.json'))
+
+    def test_ref_props_non_dict_handling(self):
+        """
+        Test added according to reported issue of 'str' object has no attribute 'items'.
+        This test runs extract outputs commands with properties that are not dict, and expects to pass.
+
+        Given:
+        - Dict containing data.
+
+        When:
+        - Extracting outputs from the given data.
+
+        Then:
+        - Ensure extract outputs does not stop on 'str' object has no attribute 'items' error.
+        """
+        base_name = 'TestSwagger'
+        integration = OpenAPIIntegration(self.swagger_path, base_name,
+                                         '-'.join(base_name.split(' ')).lower(),
+                                         base_name.replace(' ', ''),
+                                         unique_keys='id',
+                                         root_objects='Pet',
+                                         fix_code=True)
+        integration.reference = {'example': {'properties': 'type: object'}, 'example2': {'properties': {'data': 2}}}
+        data = {
+            'responses': {
+                '200': {
+                    'description': 'Response number one',
+                    'schema': [
+                        {
+                            '$ref': '#/ref/example'
+                        }
+                    ]
+                }
+            }
+        }
+        integration.extract_outputs(data)