+++ demisto_sdk/__main__.py
@@ -3,6 +3,7 @@ import json
 import logging
 import os
 import sys
+import tempfile
 from configparser import ConfigParser, MissingSectionHeaderError
 from pathlib import Path
 from typing import IO
@@ -65,6 +66,8 @@ from demisto_sdk.commands.update_release_notes.update_rn_manager import \
     UpdateReleaseNotesManager
 from demisto_sdk.commands.upload.uploader import Uploader
 from demisto_sdk.commands.validate.validate_manager import ValidateManager
+from demisto_sdk.commands.zip_packs.packs_zipper import (EX_FAIL, EX_SUCCESS,
+                                                         PacksZipper)
 
 
 class PathsParamType(click.Path):
@@ -283,6 +286,38 @@ def unify(**kwargs):
     unifier = Unifier(**kwargs)
     unifier.merge_script_package_to_yml()
     return 0
+
+
+# ====================== zip-packs ====================== #
+@main.command(hidden=True)
+@click.help_option(
+    '-h', '--help'
+)
+@click.option('-i', '--input',
+              help="The packs to be zipped as csv list of pack paths.",
+              type=PathsParamType(exists=True, resolve_path=True),
+              required=True)
+@click.option('-o', '--output', help='The destination directory to create the packs.',
+              type=click.Path(file_okay=False, resolve_path=True), required=True)
+@click.option('-v', '--content-version', help='The content version in CommonServerPython.', default='0.0.0')
+@click.option('-u', '--upload', is_flag=True, help='Upload the unified packs to the marketplace.', default=False)
+@click.option('--zip-all', is_flag=True, help='Zip all the packs in one zip file.', default=False)
+def zip_packs(**kwargs) -> int:
+    """Generating zipped packs that are ready to be uploaded to Cortex XSOAR machine."""
+    logging_setup(3)
+    check_configuration_file('zip-packs', kwargs)
+
+    # if upload is true - all zip packs will be compressed to one zip file
+    should_upload = kwargs.pop('upload', False)
+    zip_all = kwargs.pop('zip_all', False) or should_upload
+
+    packs_zipper = PacksZipper(zip_all=zip_all, pack_paths=kwargs.pop('input'), quiet_mode=zip_all, **kwargs)
+    zip_path, unified_pack_names = packs_zipper.zip_packs()
+
+    if should_upload and zip_path:
+        return Uploader(input=zip_path, pack_names=unified_pack_names).upload()
+
+    return EX_SUCCESS if zip_path is not None else EX_FAIL
 
 
 # ====================== validate ====================== #
@@ -658,6 +693,7 @@ def format(
 )
 @click.option(
     "-i", "--input",
+    type=PathsParamType(exists=True, resolve_path=True),
     help="The path of file or a directory to upload. The following are supported:\n"
          "- Pack\n"
          "- A content entity directory that is inside a pack. For example: an Integrations "
@@ -665,6 +701,13 @@ def format(
          "- Valid file that can be imported to Cortex XSOAR manually. For example a playbook: "
          "helloWorld.yml", required=True
 )
+@click.option(
+    "-z", "--zip",
+    help="Compress the pack to zip before upload, this flag is relevant only for packs.", is_flag=True
+)
+@click.option(
+    "--keep-zip", help="Directory where to store the zip after creation, this argument is relevant only for packs "
+                       "and in case the --zip flag is used.", required=False, type=click.Path(exists=True))
 @click.option(
     "--insecure",
     help="Skip certificate validation", is_flag=True
@@ -674,14 +717,27 @@ def format(
     help="Verbose output", is_flag=True
 )
 def upload(**kwargs):
-    """Upload integration to Demisto instance.
+    """Upload integration or pack to Demisto instance.
     DEMISTO_BASE_URL environment variable should contain the Demisto server base URL.
     DEMISTO_API_KEY environment variable should contain a valid Demisto API Key.
     * Note: Uploading classifiers to Cortex XSOAR is available from version 6.0.0 and up. *
     """
+    if kwargs.pop('zip', False):
+        pack_path = kwargs['input']
+        output_zip_path = kwargs.pop('keep_zip') or tempfile.gettempdir()
+        packs_unifier = PacksZipper(pack_paths=pack_path, output=output_zip_path,
+                                    content_version='0.0.0', zip_all=True, quiet_mode=True)
+        packs_zip_path, pack_names = packs_unifier.zip_packs()
+        if packs_zip_path is None:
+            return EX_FAIL
+
+        kwargs['input'] = packs_zip_path
+        kwargs['pack_names'] = pack_names
+    else:
+        kwargs.pop('keep_zip')
+
     check_configuration_file('upload', kwargs)
-    uploader = Uploader(**kwargs)
-    return uploader.upload()
+    return Uploader(**kwargs).upload()
 
 
 # ====================== download ====================== #
@@ -1135,6 +1191,9 @@ def update_release_notes(**kwargs):
 
 # ====================== find-dependencies ====================== #
 @main.command()
+@click.help_option(
+    '-h', '--help'
+)
 @click.option(
     "-i", "--input", help="Pack path to find dependencies. For example: Pack/HelloWorld", required=True,
     type=click.Path(exists=True, dir_okay=True))
@@ -1370,7 +1429,7 @@ def openapi_codegen(**kwargs):
         integration.save_config(integration.configuration, output_dir)
         tools.print_success(f'Created configuration file in {output_dir}')
         if not kwargs.get('use_default', False):
-            config_path = os.path.join(output_dir, f'{base_name}.json')
+            config_path = os.path.join(output_dir, f'{base_name}_config.json')
             command_to_run = f'demisto-sdk openapi-codegen -i "{input_file}" -cf "{config_path}" -n "{base_name}" ' \
                              f'-o "{output_dir}" -pr "{command_prefix}" -c "{context_path}"'
             if unique_keys:
@@ -1585,7 +1644,6 @@ def error_code(config, **kwargs):
 @main.resultcallback()
 def exit_from_program(result=0, **kwargs):
     sys.exit(result)
-
 
 # todo: add download from demisto command