+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -47,6 +47,7 @@ from demisto_sdk.tests.constants_test import (IGNORED_PNG,
                                               VALID_GENERIC_DEFINITION_PATH)
 from demisto_sdk.tests.test_files.validate_integration_test_valid_types import (
     LAYOUT, MAPPER, OLD_CLASSIFIER, REPUTATION)
+from pytest import raises
 from TestSuite.pack import Pack
 from TestSuite.playbook import Playbook
 from TestSuite.repo import Repo
@@ -65,8 +66,20 @@ class TestGenericFunctions:
         assert func(file_path)
 
     def test_get_file_exception(self):
+        """
+        Given
+        - A non supported file.
+
+        When
+        - Running get_file.
+
+        Then
+        - Ensure the function raise an error.
+        """
         path_to_here = f'{git_path()}/demisto_sdk/tests/test_files/'
-        assert get_file(json.load, os.path.join(path_to_here, 'fake_integration.yml'), ('yml', 'yaml')) == {}
+        with raises(ValueError) as e:
+            result = get_file(json.load, os.path.join(path_to_here, 'fake_integration.yml'), ('yml', 'yaml'))
+            assert result == e.value
 
     @pytest.mark.parametrize('dir_path', ['demisto_sdk', f'{git_path()}/demisto_sdk/tests/test_files'])
     def test_get_yml_paths_in_dir(self, dir_path):
@@ -1112,3 +1125,50 @@ def test_get_release_note_entries(version, expected_result):
     """
 
     assert get_release_note_entries(version) == expected_result
+
+
+def test_suppress_stdout(capsys):
+    """
+        Given:
+            - Messages to print.
+
+        When:
+            - Printing a message inside the suppress_stdout context manager.
+            - Printing message after the suppress_stdout context manager is used.
+        Then:
+            - Ensure that messages are not printed to console while suppress_stdout is enabled.
+            - Ensure that messages are printed to console when suppress_stdout is disabled.
+    """
+    print('You can see this')
+    captured = capsys.readouterr()
+    assert captured.out == 'You can see this\n'
+    with tools.suppress_stdout():
+        print('You cannot see this')
+        captured = capsys.readouterr()
+    assert captured.out == ''
+    print('And you can see this again')
+    captured = capsys.readouterr()
+    assert captured.out == 'And you can see this again\n'
+
+
+def test_suppress_stdout_exception(capsys):
+    """
+        Given:
+            - Messages to print.
+
+        When:
+            - Performing an operation which throws an exception inside the suppress_stdout context manager.
+            - Printing something after the suppress_stdout context manager is used.
+        Then:
+            - Ensure that the context manager do not not effect exception handling.
+            - Ensure that messages are printed to console when suppress_stdout is disabled.
+
+    """
+    with pytest.raises(Exception) as excinfo:
+        with tools.suppress_stdout():
+            x = 2 / 0
+    assert str(excinfo.value) == 'division by zero'
+    print('After error prints are enabled again.')
+    captured = capsys.readouterr()
+    assert captured.out == 'After error prints are enabled again.\n'
+