+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 * Enhanced the **postman-codegen** command to name all generated arguments with lower case.
+* Fixed an issue where the **find-dependencies** command miscalculated the dependencies for playbooks that use generic commands.
 * Fixed an issue where the **validate** command failed in external repositories in case the DEMISTO_SDK_GITHUB_TOKEN was not set.
 * Fixed an issue where **openapi-codegen** corrupted the swagger file by overwriting configuration to swagger file.
 * Updated the **upload** command to support uploading zipped packs to the marketplace.