+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 * Fixed an issue where **openapi-codegen** corrupted the swagger file by overwriting configuration to swagger file.
 * Updated the **upload** command to support uploading zipped packs to the marketplace.
 * Added to the **postman-codegen** command support of path variables.
+* Added validators for generic objects:
+  - Generic field validator - verify that the 'fromVersion' field is above 6.5.0, 'group' field equals 4 and 'id' field starts with the prefix
+    '_generics'.
+  - Generic type validator - verify that the 'fromVersion' field is above 6.5.0
+  - Generic module validator - verify that the 'fromVersion' field is above 6.5.0
+  - Generic definition validator - verify that the 'fromVersion' field is above 6.5.0
 
 # 1.4.4
 * When formatting incident types with Auto-Extract rules and without mode field, the **format** command will now add the user selected mode.